python_scripts/requirements.txt
docopt
psycopg2
gitpython
